Changed form class to interface for services
