+++ src/main/java/com/nullhawk/inventory/controller/ItemController.java
@@ -1,22 +1,18 @@
 package com.nullhawk.inventory.controller;
 
-import com.nullhawk.inventory.models.Item;
-import com.nullhawk.inventory.services.itemService;
+import com.nullhawk.inventory.services.itemServiceImp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @Controller
 @RequestMapping("/item")
 public class ItemController {
 
     @Autowired
-    private itemService service;
+    private itemServiceImp service;
 
-    public ItemController(itemService service) {
+    public ItemController(itemServiceImp service) {
         this.service = service;
     }
