+++ src/main/java/com/nullhawk/inventory/services/itemServiceImp.java
@@ -9,26 +9,30 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class itemService {
+public class itemServiceImp implements ItemService{
     @Autowired
     private ItemRepository itemRepository;
 
-    public itemService(ItemRepository itemRepository) {
+    public itemServiceImp(ItemRepository itemRepository) {
         this.itemRepository = itemRepository;
     }
-//
+
+    @Override
     public List<Item> getAllItems() {
         return itemRepository.findAll();
     }
 //
+    @Override
     public Item getItem(Long id) {
         return itemRepository.findById(id).orElseThrow(() -> new ItemNotFoundException("Item not found"));
     }
 //
+    @Override
     public Item createItem(Item item) {
         return itemRepository.save(item);
     }
 
+    @Override
     public Item updateItem(Long id, Item itemDetails) {
         Item item = getItem(id);
         item.setName(itemDetails.getName());
@@ -37,6 +41,7 @@ public class itemService {
         return itemRepository.save(item);
     }
 
+    @Override
     public void deleteItem(Long id) {
         Item item = getItem(id);
         itemRepository.delete(item);
